magit-section--maybe-wash: Also bind magit-insert-section--current
`magit-insert-section' binds both `magit-insert-section--current'
and `magit-insert-section--parent' to the same section, which, while
surprising, is actually correct.  The former is used by this macro
itself and by closely related functions that deal with the "current"
section being inserted.  The latter is for functions that wash output
inserted inside the body of `magit-section-insert', and split that up
into subsections.

Also binding `magit-insert-section--current' means that the lambda that
`magit-insert-section-body' creates and which may run delayed and thus
outside the `magit-insert-section' from, can assume a more consisten
environment.

Previously we assumed that the BODY of all `magit-insert-section-body'
are responsible for inserting subsections, for which we only need the
"parent".  When that is not the case (when the washing is delayed but
does not involve splitting into subsections) then the delayed code
instead needs to know about the "current" section.