magit-insert-section--finish: Avoid unnecessary work for root section
No longer loop over all text (again) on behalf of the root section.
It is unnecessary to do that because all text should already have been
propertized when looping over it on behalf on the child sections.

It is *possible* that there still exist root sections that do have a
heading and/or text after the last child, in which case we would now
fail to propertize that.

However, if that is the case, then that is something that should be
improved, regardless of this optimization.  The heading should be
replaced with information displayed header-line, and the trailing
text should in all likelihood just be removed.

In any case, do not perform unnecessary work in every buffer, just in
case some obscure buffer still benefits from that work.