git-commit: Disable auto-filling for summary line
The coloring of the overflowing characters makes sure we notice that
we have gone over the limit, and should be enough to make us consider
other options.  If we decide to go ahead and violate the limit anyway,
we should not have to manually revert the effect of the auto-filling.