magit-log-buffer-file: Support using the region in blobs again
Finally fix the regression introduced by [1: 293bf0d6c8].  The goal
there was to ignore the region in Dired buffers and that was done by
checking if the file (in this case `default-directory') is a regular
file.  Blobs were collateral damage of this approach, because they too
are not considered regular files.  Instead test what actually matters:
is the current buffer a Dired buffer?

1: 2018-04-02 293bf0d6c8d25cab3c583f921e9d19d42456e38b
   magit-log-buffer-file: Only trace changes if is a regular file