Report detailed hook runtime
When `magit-refresh-verbose' is set, execution times of different magit sections
and hooks is written to the *Messages* buffer. For hooks only the total runtime
of all functions in the hooks is reported.

In order to identify costly hooks, reporting the runtime for each function is
helpful. By wrapping each function of a hook, we can report individual run
times.