magit-confirm: Directly support two rounds of prompt formatting
The PROMPT and PROMPT-N arguments are used as f-strings passed to
`format'.  Many callers of this function themselves use `format'
themselves to prepare these arguments.  The "%s" in the original
f-string, which isn't to be replaced until the second round, has
to be escaped as "%%s".

Many of the first round `format' calls are used to embed branch
names and other refs into the prompt, and here a problem arises:
"%" is allowed in Git refs, so we may end up with unescaped "%d"
in the final f-string.

To avoid having to escape "%s" in many callers, add support for
handling the first round of prompt formatting inside `magit-confirm'.
PROMPT and PROMPT-N can now be lists of the form (F-STRING . ARGS),
in which case they applied to `format', after escaping every "%s" in
every argument that is a string.

Closes #5178.