magit-add-section-hook: Autoload
This makes it possible to use

  (use-package magit
    :defer t
    :config (magit-add-section-hook ...))

which essentially expands to

  (with-eval-after-load 'magit
    (magit-add-section-hook ...))

without getting a warning about `magit-add-section-hook'
possibly not being defined at runtime.

In the absence of an official autoload, users had to use

  (use-package magit
    :defer t
    :functions (magit-add-section-hook)
    :config (magit-add-section-hook ...))

but that's confusing because `:functions' only has an effect when
*compiling* a file.  When using `pp-macroexpand-last-sexp', then
we are not compiling, so there is no need to silence the compiler
by using `declare-function', and `use-package' optimizes it away,
confusing anyone who hopes macro-expansion reveals the whole truth,
but for that, one also has to read `use-package-normalize-keywords'.