magit-worktree-branch: Remove dangerous prefix argument
`magit-branch-and-checkout' doesn't support resetting a branch
while checking it out either.  `magit-worktree-branch' is more
specialized, so it makes little sense for it to provide this
footgun.

This prefix argument was not even documented (making it even more
dangerous, and considerably less likely to be used by someone who
knows what it does).