magit-diff-working-tree: Change diff type to committed
magit-diff-range with a single revision and magit-diff-working-tree
both show a diff between the revision and the working tree.  Whether
magit-reverse aborted in these buffers used to be determined by
whether the specified revision resolved to HEAD.  If so, magit-reverse
aborted with "Cannot reverse unstaged changes".

magit-reverse aborting here wasn't entirely right because a diff of
HEAD versus the working tree may contain staged changes too.

e94b6ebf (Record diff-type in magit-diff-mode buffers, 2023-03-18)
moved magit-diff-range and magit-diff-working-tree in different
directions.  With that change, magit-diff-range with any revision,
including HEAD, allows reversing.  On the other hand,
magit-diff-working-tree with any revision forbids reversing.

Update magit-diff-working-tree to pass `committed' instead of
`unstaged' for the diff type.  As a description, `committed' is a bit
off (the diff can include a mix of unstaged, staged, and committed
changes), but it allows reversing and matches what magit-diff-range
does.

For a diff of any revision against the working tree, magit-apply is
bound to fail.  Before e94b6ebf, magit-apply used to abort when the
diff was between _HEAD_ and the working tree.  After e94b6ebf and this
change, magit-apply leaves the failure to git-apply.  To abort early,
we could mark working tree diffs with a distinct diff type, but
leaving the failure to apply time is good enough, at least for now.

Closes #5090.