Add changelog file
In the past we used a separate file for each release, but since
I intend to release more often now, a single file might be more
appropriate.

This also has the benefit that the list of current changes is only
a single click away at https://github.com/magit/magit.

I might still move this to docs/RelNotes/N.org at release time and
recreate this file for round N+1.