magit-get{,-all,-boolean}: Get values from included files
When `magit--refresh-cache' is non-nil, a different code-path is taken,
which did already respect values set in included files.  But when not
using a cache, then we have to use "--include".  This only matters when
also using arguments such as "--global" and "--file".  It seems like a
bad default that git defaults to "--no-include" when such an argument
is used.