manual: Add workaround for too old Emacs on GNU ELPA build server
Emacs (and thus Org) on GNU ELPA build server is too old, so we have
to allow Org to continue when it encounters broken links.  Users will
notice these broken links, but this way we at least get to distribute
the Magit manual.

The broken links are those that link to Git's man pages.