magit-insert-recent-commits: Only use explicit range with --graph
Since [1: 55a61d15e8] we use an explicit range "HEAD~<N>..HEAD" in
addition to "-n<N>", for a massive speedup when using "--graph".

When not using "--graph", then that actually results in a (smaller)
slowdown, so stop doing it in that case.

When `core.commitGraph' is enabled (the default) and an up-to-date
commit-graph file exists (it usually does not), then we could always
omit the range, but it wouldn't make much of a difference in that case.

  core.commitGraph=false               -n10                      0.557
  core.commitGraph=false                    HEAD~10..HEAD        1.613
  core.commitGraph=false               -n10 HEAD~10..HEAD        1.569
  core.commitGraph=false    --graph    -n10                      63.776
  core.commitGraph=false    --graph         HEAD~10..HEAD        1.555
  core.commitGraph=false    --graph    -n10 HEAD~10..HEAD        1.500

  core.commitGraph=true                -n10                      0.545
  core.commitGraph=true                     HEAD~10..HEAD        0.688
  core.commitGraph=true                -n10 HEAD~10..HEAD        0.630
  core.commitGraph=true     --graph    -n10                      0.542
  core.commitGraph=true     --graph         HEAD~10..HEAD        0.598
  core.commitGraph=true     --graph    -n10 HEAD~10..HEAD        0.491

Also see #5075.

1: 2015-09-07 55a61d15e85fe46409e512cc60815e61b4843066
   avoid `git log --graph' inefficiency