magit-auto-revert-mode: Respect auto-revert-remote-files up front
This has the advantage that it avoids unnecessary work when connecting
to a remote and when later opening additional remote files, when this
is disabled, which it is by default.

The disadvantage is that this is inconsistent with the behavior of
`global-auto-revert-mode', which enables the local mode for buffers
visiting remote files regardless of the value of this variable, and
leaving it to `auto-revert-handler' to possibly forgo the process of
reverting buffers.

Closes #5222.
Closes #5223.