magit-revert-rev-file-buffer: Avoid Eglot
Eglot may be activated for buffers visiting blobs like lsp-mode, and
most LSP servers bail out in that case. So in this commit we remove the
activation function from the relevant hook.

Note that when started, Eglot checks and attempts to manage existing
buffers as well. However, at that point, buffers visiting blobs have
'buffer-file-name' being nil, and Eglot will ignore them. So we only
need to modify the hook instead of advising the function.

This is to achieve the same effect as the advice for 'lsp', introduced
in [1: f8353f57].

1: 2023-08-21 f8353f575e9ed968715a685d69501e5c105f2aee
   magit-revert-rev-file-buffer: Run major-mode hooks again