No longer use cl-callf
This macro is interesting and tt some point I might start using it
regularly, but until that day has come, I encounter it so infrequently
that I always have macroexpand, to be sure what is going on.

I also suspect that I would find this more intuitive if FUNC were
evaluated, i.e., if the value passed for that argument were usually
function-quoted.  That would make it much easier to know which argument
serves which purpose.