Better support inserting section into multiple buffers
Normally buffers are refreshed one by one.  However, the sections
may be inserted into the process buffer while another buffer is
being refreshed, in particular if `magit-git-debug' is enabled.

This lead to errors because `magit-insert-headers' which is used to
create the header sections in the status buffer adds a function to
the global value `magit-insert-section-hook'.  If we switch to the
inserting a section into the process buffer while this is happening,
that hook function adds that section to the list of sections that
are later in `magit-insert-heading' processed again.

This can be prevented by only adding that function to the local value
of the hook.

That is enough to address this particular issues, but also add some
additional safety measures.

In `magit-process-insert-section' bind `magit-insert-section--current'
to nil.  Currently this is not necessary, but future changes could make
it necessary, and since it also conceptually makes sense to bind these
when the other `magit-insert-section--*' functions are let-bound, there
is no harm in doing this proactively.

Make `magit-insert-section--current', `magit-insert-section--parent'
and `magit-insert-section--oldroot' automatically buffer local.  These
variables are always about the sections in just one particular buffer.
As long as sections are only ever inserted into one buffer at a time,
it does not make a difference if these are local or global, but since
this is already not the case (due to additions to process buffer) and
we might in the future refresh buffers asynchronously, we need buffer-
local values.