magit-process-insert-section: Improve file-path comparision
Previously this check sometimes failed due to one path ending with
a slash and the other not doing so, resulting in a "relative path"
of "." needlessly being displayed.

Don't go as far as using `file-equal-p' because that would be more
expensive and while we should minimize the chances of that happening
(as we have just done) it is okay to fail sometimes.