magit-stash-{pop,apply}: Stop after successfully installing conflict
If we cannot successfully install the conflict with "git stash apply
--index", we also try "git stash apply", and if that fails as well,
we continue with "git apply --3way".

That assumed that we could tell whether a stash was either applied
with no conflict or the conflict was successfully installed by, or
else the conflict could not be installed, by checking status.

Unfortunately it turns out that the exit status is also 1 if the
conflict was successfully installed.  Instead we have to look at
the output.

Unfortunately we can only do that for Git v2.38.0 and later, and
have to fall back to the less convenient behavior provided by Git
itself.

Closes #5253.