Require magit-process for the benefit of magit-auto-revert-mode
To avoid circular dependencies, `magit-git' does not require
`magit-process', even though it uses functions from that.  Normally
that works fine because using pretty much every feature of Magit,
causes both of these libraries to be loaded.

However, because it is autoloaded, a user may call
`magit-auto-revert-mode' in their init file, without loading Magit
first (or doing anything that triggers Magit to be loaded).  In this
case `magit-autorevert' is loaded, but not `magit', which would load
all the parts of Magit whose loading cannot be delayed.

Previously `magit-autorevert' required `magit-git', so that was
loaded too.  Now it requires `magit-process', causing all three
libraries to be loaded.

Users do not actually have to enable `magit-auto-revert-mode' in
their init file, because it is enabled by default.  Disabling the
mode did not cause an issue and it's valid to do that in the init
file.

Closes #5263.