Expel dependabot
After not doing anything for years, it suddenly began trying to run,
which failed because it is not one of the actions I have allowed for
this organization.  That didn't keep it from trying two dozen times
over the course of two weeks.  Good riddance.

I'm actually already keeping my action dependencies up-to-date, so I
don't need this.