Revert "No longer use buffer-local-value as a generalized variable"
This reverts commit b458db3263d11677a9d002a2a95ec86f29f32728.

It has been un-deprecated.  It was temporarily marked as obsolete
during the development of Emacs 29.1.

Using `setf' for the place is unproblematic, however, using it with
`cl-letf' leads to surprising results -- but that's not what we are
doing here.

Re (debbugs-gnu-bugs 26624 65555).
Re https://lists.gnu.org/archive/html/emacs-devel/2022-11/msg01406.html.