Set tabulated-list-revert-hook using setq-local
It is weird that this is a hook.  Usually this is a single function,
which refreshes the entries in the buffer.  But because it is a hook,
user could theoretically add additional functions to it, which they
wish run at the same time.

Since the previous commit, `magit-submodule-list-mode' derives from
`magit-repolist-mode', and due to the use of `add-hook' in both mode
functions, we ended up installing two refresh buffers, one of which
is not appropriate.

We could using `remove-hook' and then `add-hook', but it seems cleaner
to use `setq-local'.  We make t part of the hook's local value, so any
functions the user added globally are still run.  If a user wants to
add additional functions locally, the should probably do that using
`magit-submodule-list-mode-hook'.