git-commit.el: Require magit-process
There is a circular dependency between `magit-git' and `magit-process',
which we break by declaring some functions from the latter in the
former.  This usually works out fine because loading `magit' loads both
libraries.

However, when someone only loads `git-commit', then this broke and to
fix it we have to explicitly require `magit-process' in that library
as well.

Also explicitly require `magit-git', even though that isn't strictly
necessary because `magit-mode' requires `magit-git'.  Finally also
update the dependency declaration in "lisp/Makefile".

Closes #5280.