magit-commit-squash-internal: Always add either --no-edit or --edit
Adding "--edit" is not actually necessary, because the variants that
need it, use other arguments, which imply it.  But users don't always
know which arguments imply which of these two arguments, so it is a nice
touch to include this information in the arguments, shown in the process
buffer.  This also makes it easier to document the behavior of variants;
we can now say that some of them use "--edit", without that not actually
being true.