magit-branch-at-point: Cast more section types
Support `unpulled', `unpushed' and `pullreq' sections, like
`magit-branch-or-commit-at-point' already does.  Unlike that function,
only return an actual branch for a `pullreq' section, without falling
back to "refs/pullreqs/*"; and do not return tags.

One set of commands that benefits from this are the reset commands.
A convenient way to reset to the upstream branch now is to place the
cursor on "Unmerged into {UPSTREAM}" and typing "C-u x RET".