magit-process-yes-or-no-prompt-regexp: Fix regexp for cli passphrases
When we stopped matching text that might identify the key
in [1: a43c1d8d7a], the intention was only to avoid false-negatives
by being to restrictive.  However, since `magit-process-match-prompt'
only returns the matched text (i.e., it doesn't extend to the end of
the line), we do have to match the full line.  Allow anything after
"passphrase" and only insist on the trailing colon.

See #5291.

1: 2025-01-25 a43c1d8d7a99dd2b9f865285bed442dddb0e5ce8
   magit-process-password-prompt-regexps: Increase flexibility