magit-{diff,log}{,-refresh}: Pick up all buffer arguments
When the value of a diff or log menu was being initialized from the
arguments in the current buffer and the diff/log was already limited
to a set of files, then all other arguments were discarded.

The value of the prefix's `value' slot is a flat list of arguments
and in that list the "--" argument is represented as a list beginning
with "--", followed by the files.

This is different from `magit-{diff,log}-arguments', which returns
(ARGS FILES), which is probably how this bug came to be (but note that
the order is also different and that FILES is just a list of files,
without a "--" at its head.

Closes #5304.