magit-git-mergetool: Fix suffix invocation edge-case
Iff the user invoked the self-suffix with a prefix argument (which is
not a reasonable thing to do in this case), then the command was again
invoked as a menu.  We tried to protect against that user error, but
failed to do so, because we checked the value of the wrong variable.