Call magit-insert-heading less often without arguments
Only call `magit-insert-heading' without any arguments while washing the
output of `git-diff' or `git-log' in place, to *mark* the border between
the heading and body and change the appearance of the heading.

In all other cases, when explicitly inserting the text to be used as the
heading, use `magit-insert-heading' itself to do it, instead of `insert'
followed by `magit-insert-heading' without arguments.  We already did
this in most cases, now we consistently do it in all cases.

In the future we might instead want to use two distinct functions.