magit-thingatpt--git-revision: Only recognize painted references
In [1: a07b46fbe3] we started to ignore text that matches the name of
a reference but is propertized using a face that does not appear in
`magit-revision-faces', but if the text is not propertized, then we
continued to treated it as a revision or reference.  Now we insist on
the use of a revision or reference face.  Users who want to revert to
the old behavior can add nil to `magit-revision-faces'.

Closes #4115.

1: 2022-06-03 a07b46fbe3bf3f4737be812c4bd42a85db012a9a
   magit-thingatpt--git-revision: Avoid some false-positives