Symbol-quote booleans in more docstrings
This was forbidden until recently, and I'll need some time to develop
my personal preference, now that the dogma is gone.  For now I symbol-
quote booleans in docstrings, if they appear in a list of symbols,
which include non-booleans.  For local consistency, I quote all
appearances within a docstring, not just those in the list.

Don't quote "nil" in "non-nil".  And if booleans appear in a docstring
but never within some kind of listing, then I don't quote them either.

This is a follow up to [1: ea811b143b].

1: 2025-02-25 ea811b143b41a217cffe5d980d150437dddc4cef
   Symbol-quote booleans even in docstrings