magit-commit: Add magit-rebase-reword-commit but hide it by default
This command is also available from the `magit-rebase' menu, and has
been for a long time.

In the context of `magit-commit' it is the missing "Instant Reword"
variant, which one might expect to find here, accompanying "Instant
Fixup" and "Instant Squash".

Do not make it available by default because it is different from all
other commands in this menu, in that it does not create a commit in the
presence which is send (whether instantly or later) into the past.  In
contrast, this command has you immediately hop into a time machine and
only once you have arrived at your destiny, can you start messing with
the timeline.

While this command is unlike its second home neighbors, one can tell
that "Rebase to modify/reword/remove a commit" all grew up in the same
neighborhood, despite their diverging interests.