magit-insert-section-body: Set section properties on delayed insertion
Often when `magit-insert-section-body' is used, the section body
consists of a list of child section, followed by an empty line
belonging to the list section.  On delayed insertion of the body
we did not set the properties of that empty line.

Split out `magit--set-section-properties' from
`magit-insert-section--finish', so that we can
use it in `magit-insert-section-body' as well.