magit-insert-section--finish: Always add heading keymap
Previously we only did so if the section had children.  That was wrong
because the section may have a body that doesn't belong to children, or
it may have a `washer', which adds a body when the section is expanded.

Instead of merely also adding the heading keymap in those cases as well,
add it always.  Currently that keymap by default only features commands
to expand/collapse the section, but it is valid for someone to add more
commands.  Some users/packages might already be doing that and we might
start to do so by default.