magit-completing-read: Support empty choice when match is required
If "" is a member of COLLECTION, then do not error if the user exits
with empty input, even when REQUIRE-MATCH is t.  In that case, the
empty string/input is one of the valid matches.  Return nil instead.

This is used by `forge-read-topic-milestone' to support removing the
currently set milestone.