magit-merge-into: Rename to magit-merge-dissolve and change binding
Use the term "dissolve" because that makes it clear that the branch that
is being merged is also deleted afterwards.  Also change the key binding
in the `magit-merge' menu, to keep it mnemonic.  At least the latter was
planned since [1: 4c096921fd].  Do not define `magit-merge-into' as an
obsolete alias, since that would prolong the period during which some
users would by take by surprise.

Re #4386.

1: 2021-07-01 4c096921fda97ebdff656a35a257910fca9e5aad
   magit-merge: Change magit-merge-into's description