magit-push: Add --follow-tags and --tags
Previously users were expected to push tags explicitly either
individually (using "T"/`magit-push-tag') or all at once (using
"t"/`magit-push-tags').

The goal of this design was to avoid pushing tags only intended for
local use.  The actual result is that many users often forget to push
tags.  The fact that the release tags of many Emacs packages lag
behind the Package-Requires headers is a strong indication of that.

Git provides a better mechanism to ensure "public" tags are always
pushed:

--follow-tags
    Push all the refs that would be pushed without this option, and
    also push annotated tags in refs/tags that are missing from the
    remote but are pointing at commit-ish that are reachable from the
    refs being pushed. This can also be specified with configuration
    variable push.followTags. For more information, see push.followTags
    in git-config(1).

To take full advantage of that, one should set `push.followTags' to
true and then *always* annotate "public" tags and *never* annotate
"private" tags.

Of course not everyone knows that, so we stop short of enabling
`--follow-tags' by default (but if you see this, I strongly recommend
that you enable it locally).  Instead we make both `--follow-tags' and
`--tags' *available* by default and use verbose descriptions that spell
out the difference between the two.