magit-git-wash: Always keep error output
Remove the variable `magit--git-wash-keep-error' which previously could
be used to disable that.  The variable was added in [1: 78a979fde5] and
its value was changed to t in [2: 76253bda0d].  This has generally held
up well enough to be permanently enabled.

Function `magit--git-wash', which was added in [1: 78a979fde5], can be
used in cases where the section should be canceled in case of an error.

Also extend the docstrings of these two functions.

1: 2023-03-18 78a979fde52815242b165f083d171259db28e0b4
   Support displaying errors that occur when washing a section

2: 2024-11-15 76253bda0d05d252e118e56955447b6e4d6c0535
   magit--git-wash-keep-error: Enable by default