magit-section-make-overlay: Remove kludge for bad user setting
We strongly discourage setting `magit-section-keep-region-overlay'
to t.  If a user did that anyway that can lead to conflicts between
the involved faces.  Before we tried to automatically avoid such a
potential conflict by only using the `:foreground' of the
`magit-*-heading-selection' faces, which is very confusing for users
who want to enable this and also use another face attribute.