magit-region-sections: Prevent flashing the selection overlay here
This improves upon [1: 2d5ad5d2f9].  When clicking on a diff related
list heading, that section is now immediately highlighted as the current
section (we only want to prevent highlighting as selected).  In practice
the difference is barely noticeable, unless one clicks and holds,
without moving.

The second call to `mouse-drag-region' in `magit-diff-highlight-list' is
still necessary to ensure child sections are immediately highlighted as
current.

1: 2017-11-09 2d5ad5d2f9d9176ee1c17617080bef5dd6c6bf64
   Don't flash selection on mouse click