Finalize inserted sections after moving to the current section
After refreshing a buffer, and thus (re-)creating all sections,
`magit-section-show' has to be called on the root section to set the
correct visibility of all sections.  That function also highlights
the current section.  But when called at this time, point is at the
beginning of the buffer, so the wrong section gets highlighted.

Then, back in `magit-refresh-buffer' point is moved to the correct
section.  Finally `magit-section-update-highlight' is called from
`post-command-hook' (via `magit-section-post-command-hook') to
highlight the current section, as it happens after every command.

That function has logic to avoid needlessly re-highlighting a
section, but because the current section has actually changed (since
the highlighting that occurred during the refresh), there's nothing
that could be avoided.

For some sections (un-)highlighting can be expensive, so we want to
avoid highlighting twice.  We accomplish that by moving the call to
`magit-section-show' into `magit-refresh-buffer', after the point
adjustment.  Now the call to `magit-section-update-highlight' can
avoid (un-)highlighting because the sections that are supposed to
be highlighted were already highlighted during the refresh phase.