Use semantically correct conditional
I try to always use the semantically correct conditional.  This is
particularly important in cases like this where both meanings apply:
(1) the value is relevant and (2) there is a side-effect.  By not
using `unless' or `when' we communicate (1) and `progn' hints at (2).
The fact that it looks a bit odd this way is a feature, not a deficit.

(I realize that there are two places in existing code that should be
adjusted accordingly.)