magit--insert-related-refs: Use a single section
Use one section per category, instead of one section per ref on the
line.  This is the only place where we violated the rule that there
is only ever one section per line.

I always found it annoying that moving over these headers took more
key presses than expected when using `magit-next-section', and I
never developed an intuition for not using that command here, because
it so convincingly looks like using that is efficient enough.

An upcoming refactoring to section highlighting would break for these
sections, so let's address it now.

It is also simply unnecessary to use a section per branch.  Nowadays
commands that act on the branch at point, need no section, a painted
ref will do.  A small tweak to `magit-branch-at-point' will do.