magit-rebase--todo: Don't abbreviate empty set of commits
When rebase stops at the last commit, then the list of commits yet
to be rebases is (obviously) empty.

Don't call "git log --no-walk --format=%h [<commit>...]" in that
case, because if that is called without any revision arguments,
it shows information for HEAD.  That resulted in an error because
we later assert that the commit and abbreviation lists have the
same length.

Re #5365.