magit-make-margin-overlay: At bol set margin of previous line
That is actually what all but one caller need.

For `magit-log-wash-rev', which unlike the other callers does not
insert texts but instead moves other existing lines, it does not
matter.  Here we can just move the call to `magit-make-margin-overlay'
from before the call to `forward-line' to after that.

Do not just unconditionally default to set the margin of the previous
line, in case that doesn't work for third-party callers.  Only do so
if point is at the beginning of a line.  Also for compatibility, keep
the optional PREVIOUS-LINE argument, but ignore it if not provided.
I.e., unspecified means "dwim", not "nil, therefore for current line".

Suggested-by: Martin Joerg <martin.joerg@gmail.com>