Require empty region for single-section selection
Provided single-section selection is enabled, the selection
still becomes immediately active and visualized after invoking
`set-mark-command' on a heading.  However when then moving within
that line, as opposed to the heading of another heading, then the
selection is now deactivated.

That way the region overlay is not obscured when moving within a
single heading to select some text.