Coordinate highlighting between refresh and post-command phases
1. When a buffer is refreshed, section highlighting has to be updated.
2. More than one buffer may be refreshed at the same time and the
   current buffer may or may not be among them.
3. Therefore highlighting has to happen during the refresh phase.
4. However, re-highlighting can also be necessary when not refresh
   took place (or at least the current buffer wasn't refreshed.
5. Therefore re-highlighting has to happen during the post-command
   phase, which happens after every command in a Magit buffer.

We may therefore end up refreshing the same buffer twice.  To avoid that
record when a buffer is refreshed and during the post-command phase skip
highlighting such buffers before zpping the list of refreshed buffers.

A weaker attempt to accomplish this was already in place, but checking
whether `this-command' is a refresh command does not cast a wide enough
net.

Instead of recording buffers that have been refreshed and highlighted,
we could not highlight during the refresh phase and instead record
buffers as needing to be highlighted during the post-command phase.