Finalize inserted sections after moving to the current section
After refreshing a buffer, and thus (re-)creating all sections,
`magit-section-show' has to be called on the root section to set the
correct visibility of all sections.  That function may also highlight
the current section.  But when called at this time, point is at the
beginning of the buffer, so the wrong section gets highlighted.

Then, back in `magit-refresh-buffer' point is moved to the correct
section, before `magit-section-update-highlight' is called, and since
a different section is current now, highlighting has to be updated.

Avoiding to highlight twice was the motivation for this change, but a
later commit prevents it differently.  Despite that making this change
not strictly necessary, it feels right to move the `magit-section-show'
call to `magit-refresh-buffer'.