Call magit-status-goto-initial-section before highlighting
`magit-post-create-buffer-hook' is called right after
`magit-refresh-buffer' returned, i.e., after that called
`magit-section-update-highlight'.

Move `magit-status-goto-initial-section' from that hook to the new
`magit--initial-section-hook', which `magit-refresh-buffer' calls,
but only when the buffer is being created, not merely refreshed.

This avoids highlighting the wrong section before highlighting the
correct one.

Using such a hook is a bit of a hack.  It might be better to instead
add an option that maps modes to initial sections, instead of a hook,
but let's not commit to that just yet.