Replace magit-section-highlight-hook with two boolean options
Replace the hook `magit-section-highlight-hook' with a
generic function `magit-section-highlight'.

To preserve the ability to disable highlighting of the current
section and/or section selection highlighting, add new options,
`magit-section-highlight-current' and
`magit-section-highlight-selection'.

Do not replace the related `magit-section-unhighlight-hook' just
yet.  In two commits that will also be replace with a generic
function (though in that case its not an exact replacement).