Generalize selective section highlighting
Most sections are highlighted using a single overlay that covers the
complete section, including child sections, if any.  Nearly as simple
to highlight are sections that use two overlays, one for its own
heading and the other for its body (again covering all children).

Selective highlighting is necessary when the headings of child sections
ought to like different from their body, or when their bodies use faces
that set the background and that should not be shadowed by highlighting.

Until recently both was only needed for diff-related sections, the
later only for hunks.  So the implementation was also diff-specific
and treated this problem as a collection of special cases.

Add a new implementation that can be used by other sections as well.

The section slots `heading-highlight-face' already existed, but some
sections did not used it and instead propertized their headings using
custom code.  Now this slot should be used, as that makes it possible to
detect when selective highlighting is necessary.  If that is undesirable
for some reason the new `selective-highlight' can be set to a non-nil
value instead.

The new `magit-section-selective-highlight-p' function is used to detect
the need for selective highlighting.  Its implementation should be
trivial to understand.

For consistency add a new slot `heading-selection-face', similar to
`heading-highlight-face'.  Note that the value of this option does
not influence the decision whether selective highlighting is needed.

Add a third new slot `painted', which is unbound by default.  If it is
bound to any value, that indicates that the section has to be painted.
Starting with the next commit the actual value will indicated whether
it has already been painted, and if so, what kind of paint was used.